+++ e2e/support/helpers/index.js
@@ -5,7 +5,6 @@ export * from "./e2e-bi-basics-helpers";
 export * from "./e2e-boolean-helpers";
 export * from "./e2e-cloud-helpers";
 export * from "./e2e-collection-helpers";
-export * from "./e2e-collection-helpers-ts";
 export * from "./e2e-command-palette-helpers";
 export * from "./e2e-custom-column-helpers";
 export * from "./e2e-dashboard-helpers";
